- don't copy too much memory in incr_intervals_worker()
git-svn-id: svn://svn.code.sf.net/p/gwyddion/code@14950 25697fe4-d644-49b8-9d30-be5465f2adf3

+++ branches/GWYDDION-3/libgwy/tests/master.c
@@ -369,7 +369,7 @@ incr_intervals_worker(gpointer taskp,
 
     // Make failure much more probable if blocking does not work by creating
     // a time window between read and write.
-    guint *d = g_memdup(task->data + pos, len*sizeof(gdouble));
+    guint *d = g_memdup(task->data + pos, len*sizeof(guint));
     for (guint i = 0; i < len; i++)
         d[i]++;
     gwy_assign(task->data + pos, d, len);
